Refactor wasix http bindings
* Move wasmer bindings from wasi-types to wasi crate
  It makes much more sense to only have the host bindings in the wasi crate.
  Otherwise it comlicates feature gating for JS.
* Change module initialization to only initialize wasix_http bindings if
  the module requires them.
* Patch up bindings to return a Box<dyn FnOnce> initializer.
