+++ lib/wasi-types/Cargo.toml
@@ -28,7 +28,6 @@ time = "0.2"
 num_enum = "0.5.7"
 bitflags = "1.3.0"
 cfg-if = "1.0.0"
-wai-bindgen-wasmer = { version = "0.2.2", features = ["tracing"]}
 anyhow = "1.0.66"
 
 [dev-dependencies.pretty_assertions]
