+++ lib/wasi/Cargo.toml
@@ -59,6 +59,7 @@ wasmer-compiler-llvm = { version = "3.0.0-beta", path = "../compiler-llvm", opti
 wasmer-compiler-singlepass = { version = "3.0.0-beta", path = "../compiler-singlepass", optional = true }
 wasmer-compiler = { version = "3.0.0-beta", path = "../compiler", features = [ "translator" ], optional = true }
 http = "0.2.8"
+wai-bindgen-wasmer = { version = "0.2.2", features = ["tracing"] }
 
 [dependencies.reqwest]
 version = "0.11"
@@ -92,7 +93,7 @@ webc_runner = ["serde_cbor", "wasmer/compiler", "wasmer/cranelift"]
 webc_runner_rt_emscripten = ["wasmer-emscripten"]
 webc_runner_rt_wasi = []
 
-sys = ["wasmer/sys", "wasmer-wasi-types/sys", "webc/mmap"]
+sys = ["wasmer/sys", "wasmer-wasi-types/sys", "webc/mmap" ]
 sys-default = ["wasmer/wat", "wasmer/compiler", "sys", "logging", "host-fs", "sys-poll", "sys-thread", "host-vnet", "host-threads", "host-reqwest" ]
 sys-poll = []
 sys-thread = ["tokio/rt", "tokio/time", "tokio/rt-multi-thread"]
