+++ lib/wasi/src/bindings/mod.rs
@@ -1,1 +1,3 @@
-pub mod http;
+// TODO: remove allow()
+#[allow(dead_code)]
+pub mod wasix_http_client_v1;
