+++ lib/wasi/src/syscalls/wasix/proc_fork.rs
@@ -180,7 +180,7 @@ pub fn proc_fork<M: MemorySize>(
                     let mut ctx = WasiFunctionEnv::new(&mut store, child_env);
 
                     // Let's instantiate the module with the imports.
-                    let (mut import_object, init) = import_object_for_all_wasi_versions(&mut store, &ctx.env);
+                    let (mut import_object, init) = import_object_for_all_wasi_versions(&module, &mut store, &ctx.env);
                     let memory = if let Some(memory) = memory {
                         let memory = Memory::new_from_existing(&mut store, memory);
                         import_object.define("env", "memory", memory.clone());
